audio: src: move out src component data out from src source
move component data out of source file to avoid ipc3 and ipc4
usage in src source file.
